requirements.txt
numpy
PyYAML
termcolor
torch
torchinfo
torchvision
tqdm